feat: Doesn't everything really come down to relationships?!?! :smile:
I decided to model the relationships separately. Careful observation
revealed a nice, tight, circular relationship. For more relationships
we way want to employ a more comprehensive or rich model but I feel
that this can continue to work for a while.

Adding a 'cannon' would cause a change to the 'beats' property by
requiring us to use a Set. Of course, we
could move to a richer data structure to represent the circularly
linked list but there is no need right now.

We could use a DSL to model the relationship at some point.

That's it for now. I'm going on a hike....